Add CLI accession export task (#1862)
* Update accession flatfile config file

* Change creators->eventActors in sample accession CSV

Duplicate events were being created in the
database when importing accessions using the example accession CSV. This
looks to be related to the code in `csvImportBaseTask::importEvents`.
When the "creators" column is present and a column like "eventTypes" is
also present, then code duplicates the number of events because an event
is added when $version=='2.2' AND when $version=='2.3'.

Changing creators to eventActors makes it so that an event is only added
when $version=='2.3'

* Add CSV accession exporter class

* Add csv:accession-export CLI task

* Add missing column to docstring

* Set to empty string instead of NULL for missing alternative identifier fields

* Add call to parent class in modifyRowBeforeExport

---------

Co-authored-by: Kelvin Oo <kelvinoo123@gmail.com>